Fix Kubeip not replacing IPs if node's access config name !=external-nat
The steps that Kubeip follows to replace an ephemeral IP with an
external IP on a given node, among others, are:
* Delete the node's access config called 'external-nat'.
* Create a new access config for the node with a free external IP.

If the node's access config has a name different than 'external-nat', it
doesn't get deleted and, when attempting to add another access config with
the external IP, the GCloud API returns a 400:

"googleapi: Error 400: At most one access config currently supported.
badRequest"

To work this around, instead of assuming that the access config will be
called 'external-nat', actively retrieve the access config name and use
this retrieved name to issue the DeleteAccessConfig request.